# Reduce allocations in Find and ReplaceAll