- Ensure that findIter and overlappingIter do not escape to the heap in Next
- Have FindAll use a findIter directly rather than using an Iter
- Do not allocate a slice of replacements in ReplaceAll
- Right-size the output buffer in ReplaceAll